libre/requirements/common.txt
Django==1.5.1
xlrd==0.9.2
South==0.8.1
django-jsonfield==0.9.10
django-model-utils==1.4.0
pytz==2013b
suds==0.4
django-suit==0.2.2
Fiona==0.16.1
djangorestframework==2.3.6
PyYAML==3.10
pyproj==1.9.3
python-dateutil==2.1
Shapely==1.2.17
jsonpath-rw==1.2.0
Pillow==2.1.0
